Create create a currency converter in java
The purpose of this Java project is to convert a currency from one to another. 

A web-based interface for converting currency and getting the output value, for example, here displays converting the currency of the US dollar to INR. 

We see variations with different countries using different currencies. Be aware of the current exchange prices in the market and then can convert from one currency to another.